backend/requirements-render.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
openpyxl==3.1.2
python-dotenv==1.0.0
asyncpg==0.29.0
httpx==0.25.2
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# OCR and Table Parsing
pytesseract==0.3.10
Pillow==10.1.0
camelot-py[cv]==0.11.0
tabula-py==2.8.2
opencv-python-headless==4.8.1.78
# Additional dependencies for better table detection
pdf2image==1.16.3
numpy==1.24.3 